+++ packages/runtime/src/index.ts
@@ -16,11 +16,16 @@ import {
   generateCustomHeaders,
 } from './helpers/config'
 import { onPreDev } from './helpers/dev'
-import { enableEdgeInNextConfig, writeEdgeFunctions, loadMiddlewareManifest } from './helpers/edge'
+import {
+  enableEdgeInNextConfig,
+  writeEdgeFunctions,
+  loadMiddlewareManifest,
+  cleanupEdgeFunctions,
+} from './helpers/edge'
 import { moveStaticPages, movePublicFiles, patchNextFiles } from './helpers/files'
 import { generateFunctions, setupImageFunction, generatePagesResolver, getApiRouteConfigs } from './helpers/functions'
 import { generateRedirects, generateStaticRedirects } from './helpers/redirects'
-import { shouldSkip, isNextAuthInstalled, getCustomImageResponseHeaders } from './helpers/utils'
+import { shouldSkip, isNextAuthInstalled, getCustomImageResponseHeaders, getRemotePatterns } from './helpers/utils'
 import {
   verifyNetlifyBuildVersion,
   checkNextSiteHasBuilt,
@@ -73,13 +78,13 @@ const plugin: NetlifyPlugin = {
 
     checkNextSiteHasBuilt({ publish, failBuild })
 
-    let experimentalRemotePatterns = []
     const { appDir, basePath, i18n, images, target, ignore, trailingSlash, outdir, experimental } = await getNextConfig(
       {
         publish,
         failBuild,
       },
     )
+    await cleanupEdgeFunctions(constants)
 
     const middlewareManifest = await loadMiddlewareManifest(netlifyConfig)
 
@@ -109,10 +114,6 @@ const plugin: NetlifyPlugin = {
       }
     }
 
-    if (experimental.images) {
-      experimentalRemotePatterns = experimental.images.remotePatterns || []
-    }
-
     if (isNextAuthInstalled()) {
       const config = await getRequiredServerFiles(publish)
 
@@ -120,10 +121,20 @@ const plugin: NetlifyPlugin = {
 
       if (userDefinedNextAuthUrl) {
         console.log(
-          `NextAuth package detected, NEXTAUTH_URL environment variable set by user to ${userDefinedNextAuthUrl}`,
+          `NextAuth package detected, NEXTAUTH_URL environment variable set by user in next.config.js to ${userDefinedNextAuthUrl}`,
         )
+      } else if (process.env.NEXTAUTH_URL) {
+        // When the value is specified in the netlify.toml or the Netlify UI (will be evaluated in this order)
+        const nextAuthUrl = `${process.env.NEXTAUTH_URL}${basePath}`
+
+        console.log(
+          `NextAuth package detected, NEXTAUTH_URL environment variable set by user in Netlify configuration to ${nextAuthUrl}`,
+        )
+        config.config.env.NEXTAUTH_URL = nextAuthUrl
+
+        await updateRequiredServerFiles(publish, config)
       } else {
-        const nextAuthUrl = `${process.env.URL}${basePath}`
+        const nextAuthUrl = `${process.env.DEPLOY_PRIME_URL}${basePath}`
 
         console.log(`NextAuth package detected, setting NEXTAUTH_URL environment variable to ${nextAuthUrl}`)
         config.config.env.NEXTAUTH_URL = nextAuthUrl
@@ -158,7 +169,7 @@ const plugin: NetlifyPlugin = {
       imageconfig: images,
       netlifyConfig,
       basePath,
-      remotePatterns: experimentalRemotePatterns,
+      remotePatterns: getRemotePatterns(experimental, images),
       responseHeaders: getCustomImageResponseHeaders(netlifyConfig.headers),
     })