+++ packages/runtime/src/templates/getHandler.ts
@@ -49,7 +49,7 @@ const makeHandler = (conf: NextConfig, app, pageRoot, staticManifest: Array<[str
   conf.experimental.isrFlushToDisk = false
   // This is our flag that we use when patching the source
   // eslint-disable-next-line no-underscore-dangle
-  process.env._BYPASS_SSG = 'true'
+  process.env._REVALIDATE_SSG = 'true'
   for (const [key, value] of Object.entries(conf.env)) {
     process.env[key] = String(value)
   }