+++ packages/next/CHANGELOG.md
@@ -1,11 +1,25 @@
 # Changelog
 
-## [1.1.1](https://github.com/netlify/next-runtime/compare/next-v1.1.0...next-v1.1.1) (2022-08-31)
+## [1.3.0](https://github.com/netlify/next-runtime/compare/next-v1.2.0...next-v1.3.0) (2022-09-09)
 
+
+### Features
+
+* add support for Next 12.3 middleware matchers ([#1612](https://github.com/netlify/next-runtime/issues/1612)) ([fd88b98](https://github.com/netlify/next-runtime/commit/fd88b9829f93953a06ba0d1269ac8ba5f51f6874))
+
+## [1.2.0](https://github.com/netlify/next-runtime/compare/next-v1.1.1...next-v1.2.0) (2022-09-08)
+
+
+### Features
+
+* create edge function demo ([#1586](https://github.com/netlify/next-runtime/issues/1586)) ([43a60c8](https://github.com/netlify/next-runtime/commit/43a60c88260e897d3b6b49f9f05442151da36644))
+
+## [1.1.1](https://github.com/netlify/next-runtime/compare/next-v1.1.0...next-v1.1.1) (2022-08-31)
 
 ### Bug Fixes
 
-* remove Next.js as a peer dependency ([#1584](https://github.com/netlify/next-runtime/issues/1584)) ([6be4dc0](https://github.com/netlify/next-runtime/commit/6be4dc08e5339efb84e180e9ea02ce0bc6efe5b5))
+- remove Next.js as a peer dependency ([#1584](https://github.com/netlify/next-runtime/issues/1584))
+  ([6be4dc0](https://github.com/netlify/next-runtime/commit/6be4dc08e5339efb84e180e9ea02ce0bc6efe5b5))
 
 ## [1.1.0](https://github.com/netlify/next-runtime/compare/next-v1.0.0...next-v1.1.0) (2022-08-22)